requirements.txt
opencv-python~=4.8.0
numpy~=1.26.0
ffmpeg-python~=0.2.0
transformers~=4.38.0
torch~=2.6.0
pillow~=10.0.0
python-dotenv~=1.0.0
openai~=1.13.0
azure-ai-inference~=1.0.0b9
tqdm~=4.66.1
requests~=2.31.0
ultralytics~=8.1.0  # YOLO 推理支持
